Fix canvas translation
This only worked accidentally before, because yMin = -(ySize + yMin),
i.e. zero on the y-axis is exactly halfway up, but that isn’t the case
in general.

I also chose to move the “positive is up” scaling earlier, because
fixing the axes sooner makes it easier to understand the translations.

I could combine these transformations into a single one (ditto for the
scales), but keeping them separate makes it clearer what’s going on.